move status updating to a separate file
